+++ examples/oft-hyperliquid/solhint.config.js
@@ -1,1 +1,7 @@
-module.exports = require('@layerzerolabs/solhint-config');
+module.exports = {
+    ...require('@layerzerolabs/solhint-config'),
+    rules: {
+        ...require('@layerzerolabs/solhint-config').rules,
+        'contract-name-camelcase': 'off',
+    },
+};