+++ examples/oft-hyperliquid/test/foundry/HyperLiquidComposer.t.sol
@@ -5,7 +5,6 @@ import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 import { OFTComposeMsgCodec } from "@layerzerolabs/oft-evm/contracts/libs/OFTComposeMsgCodec.sol";
 
-import { IHyperAsset } from "@layerzerolabs/hyperliquid-composer/contracts/interfaces/IHyperLiquidComposerCore.sol";
 import { IHYPEPrecompile } from "@layerzerolabs/hyperliquid-composer/contracts/interfaces/IHYPEPrecompile.sol";
 import { ICoreWriter } from "@layerzerolabs/hyperliquid-composer/contracts/interfaces/ICoreWriter.sol";
 
@@ -23,24 +22,23 @@ contract HyperLiquidComposerTest is HyperliquidBaseTest {
     }
 
     function test_deployment() public view {
-        IHyperAsset memory hypeAsset = hyperLiquidComposer.getHypeAsset();
-        assertEq(hypeAsset.assetBridgeAddress, 0x2222222222222222222222222222222222222222);
+        assertEq(hyperLiquidComposer.NATIVE_ASSET_BRIDGE(), 0x2222222222222222222222222222222222222222);
         uint256 expectedHypeCoreIndexId = block.chainid == 998 ? 1105 : 150;
-        assertEq(hypeAsset.coreIndexId, expectedHypeCoreIndexId);
-        assertEq(hypeAsset.decimalDiff, 10);
+        assertEq(hyperLiquidComposer.NATIVE_CORE_INDEX_ID(), expectedHypeCoreIndexId);
+        assertEq(hyperLiquidComposer.NATIVE_DECIMAL_DIFF(), 10);
 
-        IHyperAsset memory oftAsset = hyperLiquidComposer.getOFTAsset();
-        assertEq(oftAsset.assetBridgeAddress, ERC20.assetBridgeAddress);
-        assertEq(oftAsset.coreIndexId, ERC20.coreIndexId);
-        assertEq(oftAsset.decimalDiff, ERC20.decimalDiff);
+        assertEq(hyperLiquidComposer.ERC20_ASSET_BRIDGE(), ERC20.assetBridgeAddress);
+        assertEq(hyperLiquidComposer.ERC20_CORE_INDEX_ID(), ERC20.coreIndexId);
+        assertEq(hyperLiquidComposer.ERC20_DECIMAL_DIFF(), ERC20.decimalDiff);
     }
 
     function test_hypeIndexByChainId_testnet() public {
         string memory rpcUrl = "https://rpc.hyperliquid-testnet.xyz/evm";
+
         try vm.envString("RPC_URL_HYPERLIQUID_TESTNET") returns (string memory _rpcUrl) {
             rpcUrl = _rpcUrl;
         } catch {
-            console.log("Using default mainnet RPC URL");
+            console.log("Using default testnet RPC URL");
         }
         // Skip test if fork fails
         try vm.createSelectFork(rpcUrl) {} catch {
@@ -50,13 +48,14 @@ contract HyperLiquidComposerTest is HyperliquidBaseTest {
 
         MyOFT oftTestnet = new MyOFT("test", "test", HL_LZ_ENDPOINT_V2_TESTNET, msg.sender);
         HyperLiquidComposer hypeComposerTestnet = new HyperLiquidComposer(
-            HL_LZ_ENDPOINT_V2_TESTNET,
             address(oftTestnet),
             ERC20.coreIndexId,
             ERC20.decimalDiff
         );
 
-        assertEq(hypeComposerTestnet.hypeIndexByChainId(998), 1105);
+        uint64 coreIndexId = hypeComposerTestnet.NATIVE_CORE_INDEX_ID();
+
+        assertEq(coreIndexId, 1105);
     }
 
     function test_hypeIndexByChainId_mainnet() public {
@@ -74,35 +73,42 @@ contract HyperLiquidComposerTest is HyperliquidBaseTest {
 
         MyOFT oftMainnet = new MyOFT("test", "test", HL_LZ_ENDPOINT_V2_MAINNET, msg.sender);
         HyperLiquidComposer hypeComposerMainnet = new HyperLiquidComposer(
-            HL_LZ_ENDPOINT_V2_MAINNET,
             address(oftMainnet),
             ERC20.coreIndexId,
             ERC20.decimalDiff
         );
 
-        assertEq(hypeComposerMainnet.hypeIndexByChainId(999), 150);
+        uint64 coreIndexId = hypeComposerMainnet.NATIVE_CORE_INDEX_ID();
+
+        assertEq(coreIndexId, 150);
     }
 
     function test_SendSpot_no_FundAddress() public {
-        bytes memory composeMsg = abi.encode(0, abi.encodePacked(userB));
+        bytes memory composeMsg = abi.encode(0, userB);
 
         // Build composerMsg similar to the outcome of OFTCore.send()
         bytes memory composerMsg_ = OFTComposeMsgCodec.encode(
             0,
             ETH_EID,
-            AMOUNT_TO_SEND + DUST,
+            AMOUNT_TO_SEND,
             abi.encodePacked(addressToBytes32(userA), composeMsg)
         );
 
-        deal(address(oft), address(hyperLiquidComposer), AMOUNT_TO_SEND + DUST);
+        deal(address(oft), address(hyperLiquidComposer), AMOUNT_TO_SEND);
 
-        assertEq(oft.balanceOf(address(hyperLiquidComposer)), AMOUNT_TO_SEND + DUST);
+        assertEq(oft.balanceOf(address(hyperLiquidComposer)), AMOUNT_TO_SEND);
 
         // Expect the Transfer event to be emitted
         vm.expectEmit(address(oft));
         emit IERC20.Transfer(address(hyperLiquidComposer), ERC20.assetBridgeAddress, AMOUNT_TO_SEND);
 
-        uint64 coreAmount = hyperLiquidComposer.quoteHyperCoreAmount(AMOUNT_TO_SEND, true).core;
+        uint64 coreIndexId = hyperLiquidComposer.ERC20_CORE_INDEX_ID();
+        int8 decimalDiff = hyperLiquidComposer.ERC20_DECIMAL_DIFF();
+        address assetBridgeAddress = hyperLiquidComposer.ERC20_ASSET_BRIDGE();
+
+        uint64 coreAmount = hyperLiquidComposer
+            .quoteHyperCoreAmount(coreIndexId, decimalDiff, assetBridgeAddress, AMOUNT_TO_SEND)
+            .core;
         bytes memory action = abi.encode(userB, ERC20.coreIndexId, coreAmount);
         bytes memory payload = abi.encodePacked(hyperLiquidComposer.SPOT_SEND_HEADER(), action);
         vm.expectEmit(HLP_CORE_WRITER);
@@ -115,11 +121,11 @@ contract HyperLiquidComposerTest is HyperliquidBaseTest {
         vm.stopPrank();
 
         assertEq(oft.balanceOf(address(hyperLiquidComposer)), 0);
-        assertEq(oft.balanceOf(userB), balanceBefore + DUST);
+        assertEq(oft.balanceOf(userB), balanceBefore);
     }
 
     function test_SendSpot_and_FundAddress() public {
-        bytes memory composeMsg = abi.encode(AMOUNT_TO_FUND, abi.encodePacked(userB));
+        bytes memory composeMsg = abi.encode(AMOUNT_TO_FUND, userB);
 
         // Build composerMsg similar to the outcome of OFTCore.send()
         bytes memory composerMsg_ = OFTComposeMsgCodec.encode(
@@ -132,36 +138,37 @@ contract HyperLiquidComposerTest is HyperliquidBaseTest {
         deal(address(oft), address(hyperLiquidComposer), AMOUNT_TO_SEND);
         assertEq(oft.balanceOf(address(hyperLiquidComposer)), AMOUNT_TO_SEND);
 
-        // Expect the Received event to be emitted - this is for the HYPE precompile
-        vm.expectEmit(hyperLiquidComposer.getHypeAsset().assetBridgeAddress);
-        emit IHYPEPrecompile.Received(address(hyperLiquidComposer), AMOUNT_TO_FUND);
-
         // Expect the Transfer event to be emitted
         vm.expectEmit(address(oft));
         emit IERC20.Transfer(address(hyperLiquidComposer), ERC20.assetBridgeAddress, AMOUNT_TO_SEND);
 
-        uint64 coreAmount = hyperLiquidComposer.quoteHyperCoreAmount(AMOUNT_TO_SEND, true).core;
+        uint64 coreIndexId = hyperLiquidComposer.ERC20_CORE_INDEX_ID();
+        int8 decimalDiff = hyperLiquidComposer.ERC20_DECIMAL_DIFF();
+        address assetBridgeAddress = hyperLiquidComposer.ERC20_ASSET_BRIDGE();
+
+        uint64 coreAmount = hyperLiquidComposer
+            .quoteHyperCoreAmount(coreIndexId, decimalDiff, assetBridgeAddress, AMOUNT_TO_SEND)
+            .core;
         bytes memory action = abi.encode(userB, ERC20.coreIndexId, coreAmount);
         bytes memory payload = abi.encodePacked(hyperLiquidComposer.SPOT_SEND_HEADER(), action);
         vm.expectEmit(HLP_CORE_WRITER);
         emit ICoreWriter.RawAction(address(hyperLiquidComposer), payload);
 
+        // Expect the Received event to be emitted - this is for the HYPE precompile
+        address hypeAssetBridge = hyperLiquidComposer.NATIVE_ASSET_BRIDGE();
+        vm.expectEmit(hypeAssetBridge);
+        emit IHYPEPrecompile.Received(address(hyperLiquidComposer), AMOUNT_TO_FUND);
+
         uint256 balanceBeforeBridge = HYPE.assetBridgeAddress.balance;
         uint256 balanceBeforeUserB = userB.balance;
 
         vm.startPrank(HL_LZ_ENDPOINT_V2);
-        hyperLiquidComposer.lzCompose{ value: AMOUNT_TO_FUND + DUST }(
-            address(oft),
-            bytes32(0),
-            composerMsg_,
-            msg.sender,
-            ""
-        );
+        hyperLiquidComposer.lzCompose{ value: AMOUNT_TO_FUND }(address(oft), bytes32(0), composerMsg_, msg.sender, "");
         vm.stopPrank();
 
         assertEq(oft.balanceOf(address(hyperLiquidComposer)), 0);
         assertEq(HYPE.assetBridgeAddress.balance, balanceBeforeBridge + AMOUNT_TO_FUND);
-        assertEq(userB.balance, balanceBeforeUserB + DUST);
+        assertEq(userB.balance, balanceBeforeUserB);
     }
 
     function test_getBalanceOfHyperCore(uint64 _balance) public {
@@ -171,7 +178,7 @@ contract HyperLiquidComposerTest is HyperliquidBaseTest {
             _balance
         );
 
-        uint64 balance = hyperLiquidComposer.balanceOfHyperCore(HYPE.assetBridgeAddress, HYPE.coreIndexId);
+        uint64 balance = hyperLiquidComposer.spotBalance(HYPE.assetBridgeAddress, HYPE.coreIndexId).total;
         assertEq(balance, _balance);
     }