+++ examples/native-oft-adapter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/native-oft-adapter-example
 
+## 0.2.10
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.2.9
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.2.8
 
 ### Patch Changes