+++ examples/oft-adapter-initia/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/oft-adapter-initia-example
 
+## 0.1.2
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.1.1
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.1.0
 
 ### Minor Changes