+++ examples/oft-alt/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/oft-alt-example
 
+## 0.0.5
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.0.4
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
+## 0.0.3
+
+### Patch Changes
+
+- e5bf367: feat: oapp and oft alt upgradeable
+
 ## 0.0.2
 
 ### Patch Changes