+++ examples/oft-solana/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @layerzerolabs/oft-solana-example
 
+## 0.12.2
+
+### Patch Changes
+
+- e415855: rename retryPayload to retryMessage
+- d48633e: fix oft-solana's debug script when peers are checked
+- c44e657: update evm network used
+
+## 0.12.1
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.12.0
 
 ### Minor Changes