+++ examples/oft/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/oft-example
 
+## 0.7.8
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.7.7
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.7.6
 
 ### Patch Changes