+++ packages/hyperliquid-composer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @layerzerolabs/hyperliquid-composer
 
+## 1.0.3
+
+### Patch Changes
+
+- 97f506f: update register-spot command and add in commands to track spotMarket gas and current deployment for a token
+
+## 1.0.2
+
+### Patch Changes
+
+- 90417ab: adding in freeze and quote asset during hip-1 deployment
+
+## 1.0.1
+
+### Patch Changes
+
+- 166816a: chore: add virtual to \_getFinalCoreAmount in FeeToken
+
+## 1.0.0
+
+### Major Changes
+
+- 62c0122: hyperliquid v1 with extensions for fee-token and recovery
+
 ## 0.0.19
 
 ### Patch Changes